TASK: Allow passing exceedingArguments to import:batch command
Exceeding arguments are unnamed arguments that are made available in the
DataProvider. This allows you to pass some arguments at runtime instead of
by configuration.

This is inspired by the KickstartController of Neos.Kickstart:
https://github.com/neos/kickstart/blob/master/Classes/Command/KickstartCommandController.php#L226

You must handle exceeding arguments yourself in the DataProvider. You can
get them through $this->getExceedingArguments().